# components/ambient-api-server/pkg/api/openapi/api/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: ambient-code@redhat.com
  description: Ambient API Server
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  title: Ambient API Server
  version: 1.0.0
servers:
- description: current domain
  url: http://localhost:8000
- description: Production
  url: https://api.ambient-code.com
- description: Staging
  url: https://api.stage.ambient-code.com
security:
- Bearer: []
paths:
...
  /api/ambient/v1/credentials/{cred_id}/token:
    get:
      description: Returns the decrypted token value for the given credential. Requires
        token-reader role.
      parameters:
      - description: The id of the credential
        in: path
        name: cred_id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CredentialTokenResponse"
          description: Credential token retrieved successfully
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Auth token is invalid
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Unauthorized to perform operation
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: No credential with specified id exists
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Unexpected error retrieving token
      security:
      - Bearer: []
      summary: Get a decrypted token for a credential
...
  /api/ambient/v1/projects/{id}/credentials/{cred_id}/token:
    get:
      description: Returns the decrypted token value for the given credential. Requires
        token-reader role.
      parameters:
      - description: The id of record
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      - description: The id of the credential
        in: path
        name: cred_id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CredentialTokenResponse"
          description: Credential token retrieved successfully
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Auth token is invalid
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Unauthorized to perform operation
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: No credential with specified id exists
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Unexpected error retrieving token
      security:
      - Bearer: []
      summary: Get a decrypted token for a project credential
...
components:
...
  schemas:
    ObjectReference:
      properties:
        id:
          type: string
        kind:
          type: string
        href:
          type: string
        created_at:
          format: date-time
          type: string
        updated_at:
          format: date-time
          type: string
      type: object
...
    Error:
      allOf:
      - $ref: "#/components/schemas/ObjectReference"
      - properties:
          code:
            type: string
          reason:
            type: string
          operation_id:
            type: string
        type: object
      example:
        reason: reason
        code: code
        updated_at: 2000-01-23T04:56:07.000+00:00
        kind: kind
        created_at: 2000-01-23T04:56:07.000+00:00
        operation_id: operation_id
        id: id
        href: href
...
    CredentialTokenResponse:
      example:
        provider: github
        credential_id: credential_id
        token: token
      properties:
        credential_id:
          description: ID of the credential
          type: string
        provider:
          description: Provider type for this credential
          enum:
          - github
          - gitlab
          - jira
          - google
          - vertex
          - kubeconfig
          type: string
        token:
          description: Decrypted token value
          type: string
      required:
      - credential_id
      - provider
      - token
      type: object
...
  securitySchemes:
    Bearer:
      bearerFormat: JWT
      scheme: bearer
      type: http