# components/ambient-sdk/go-sdk/types/base.go
// Code generated by ambient-sdk-generator from openapi.yaml — DO NOT EDIT.
// Source: ../../ambient-api-server/openapi/openapi.yaml
// Spec SHA256: b033e32bec23ae0f0443bfe66148d0e29b42fa6bb0223ab012bdaf2860a2858e
// Generated: 2026-07-07T12:19:25Z

package types

import (
	"strconv"
	"time"
)

type ObjectReference struct {
	ID        string     `json:"id,omitempty"`
	Kind      string     `json:"kind,omitempty"`
	Href      string     `json:"href,omitempty"`
	CreatedAt *time.Time `json:"created_at,omitempty"`
	UpdatedAt *time.Time `json:"updated_at,omitempty"`
}

type ListMeta struct {
	Kind  string `json:"kind"`
	Page  int    `json:"page"`
	Size  int    `json:"size"`
	Total int    `json:"total"`
}

type APIError struct {
	ID          string `json:"id,omitempty"`
	Kind        string `json:"kind,omitempty"`
	Href        string `json:"href,omitempty"`
	Code        string `json:"code"`
	Reason      string `json:"reason"`
	OperationID string `json:"operation_id,omitempty"`
	StatusCode  int    `json:"-"`
}

func (e *APIError) Error() string {
	if e.StatusCode > 0 {
		return "ambient API error " + strconv.Itoa(e.StatusCode) + ": " + e.Code + " — " + e.Reason
	}
	return "ambient API error: " + e.Code + " — " + e.Reason
}